dwbank/server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
billiard==4.1.0
bitarray==2.8.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coincurve==18.0.0
cytoolz==0.12.2
Django==4.2.5
django-cors-headers==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
ecdsa==0.18.0
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
h11==0.14.0
hdwallet==2.2.1
hexbytes==0.3.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
lru-dict==1.2.0
mnemonic==0.20
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.24.3
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.0.0
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
toolz==0.12.0
tronpy==0.4.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
web3==6.9.0
websockets==11.0.3
yarl==1.9.2
